min price filtering complete. Will have to reseed on pulldown

+++ bangazonapi/models/product.py
@@ -18,7 +18,9 @@ class Product(SafeDeleteModel):
     customer = models.ForeignKey(
         Customer, on_delete=models.DO_NOTHING, related_name="products"
     )
-    price = models.FloatField(
+    price = models.DecimalField(
+        max_digits=10,
+        decimal_places=2,
         validators=[MinValueValidator(0.00), MaxValueValidator(10000.00)],
     )
     description = models.CharField(
